Disabled the projection matrix translation for each eye and now just using a single projection matrix for both eyes. May need to revisit this later.
Fixed issue with stereo translation where the eyes were swapped due to mixed world and view space transformations. With this change and the change to the projection matrix calculation, stereo rendering appears to now be working!

Added model loading code and a placeholder model from Sketchup as an improvement to the axes triangles.

Camera game-translation is now view-space and not model space - fixes the issue where movement was inverted.

Q and E now modify the IPD value in meters, and added code to display the IPD value while you are modifying it.